Fix timezone issue when parsing .fit files
The .fit file contains no timezone information. Therefore allow a timezone
offset to be specified to the FitParser constructor. This offset will be
applied to all timestamps read from the .fit file in order to convert the
parsed timestamps into a Unix timestamp (seconds since _UTC_ epoch).

If no timezone offset is supplied, then derive the timezone offset as
follows: Assume the .fit file timestamps are recorded in the same time
zone as the current local time zone, and then apply the timezone offset
that were in effect at the time the .fit file were recorded (taking DST
into account).