Initial version of correlate_activities.py
This script attempts to find the offset (in #samples) between two activity
files that maximize their correlation. For now, the correlation is inversely
proportional to the average pairwise interdistance between samples in the two
samples streams (aka. files). In other words, this finds the offset to be
added to the sample index from the second file, in order to find the
corresponding sample index from the first file and on total maximizes the
correlation between the two files.