indoor_time_fixer.py: Fix timestamps in workouts where Garmin has screwy clock
Sometimes, during indoor workouts where I've set GPS OFF on the Garmin, it
will still get a GPS lock at some point, which will cause it to reset its
internal clock, typically jumping forwards a couple of days or something.
This is reflected in the resulting activity file, and you get activities
with e.g. elapsed time of 54 hours, and such...

This script fixes this in the simplest possible manner: You give it a
start timestamp, and it rewrites _all_ timestamps in the TCX file to
be in 1-second increments from that timestamp. Obviously this will
cancel/erase any _actual_ pauses you had during that workout, but most
indoor workouts are fairly simple start->finish activities in that regard.